desenvolvimento/src/requirements.txt
click==8.1.3
Flask==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.2
Wand==0.6.11
Werkzeug==2.2.3
redis
pymongo
flask-jwt-extended
pyjwt
bcrypt
flask_bcrypt